+++ setup.py
@@ -23,7 +23,7 @@ setup(
     long_description_content_type="text/markdown",
     url="https://github.com/abstra-app/abstra-cli",
     license="MIT",
-    packages=["abstra_cli"],
+    packages=["abstra_cli", "abstra_cli.resources"],
     install_requires=[
         "setuptools",
         "requests",
